refactor: move ToolResult into base_tool.py and fix circular imports
Move ToolResult from its own file into base_tool.py alongside Tool and
DeterministicToolOutput — all tool-related data types now live in one
place. Restore lazy loading of EnvironmentConfig and SynthesisConfig
in oumi.core.configs.__init__ to prevent circular imports when
oumi.environments is imported first.
