--- src/oumi/environments/tool_result.py
@@ -1,28 +1,0 @@
-# Copyright 2025 - Oumi
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""Tool execution results."""
-
-from dataclasses import dataclass
-from typing import Any
-
-from oumi.core.configs.params.base_params import BaseParams
-
-
-@dataclass
-class ToolResult(BaseParams):
-    """Result returned by an environment step."""
-
-    output: str | dict[str, Any] | None
-    updated_state: dict[str, Any] | None = None
